Move internal tests to _internal_test.go files and make some tests external
